Removed all collection sharing source
This change removes all sharing server and client functionality and all
code that it depends on.

"Collection sharing" was similar to the iTunes feature "Local Network
Sharing" from 2003, allowing to access ro collections, playlists and smart
playlists on a SIDPLAY instance running on a different Mac on the same
network. It is unclear whether the feature was ever enabled – it's disabled
in the 4.2 release.

In any case, the feature is not useful in 2025 any more, with everyone
having a full copy of HVSC, the probability of meeting someone on the same
network with SIDPLAY for Mac and different music is very low. With the
options of keeping 4000 lines of dead code, fixing up and re-enabling the
feature, or deleting it and heavily cleaning up the code base, I think we
should do the latter.